make Bullet, Enum and Structs
