+++ Jilltest.cpp
@@ -13,6 +13,12 @@ and may not be redistributed without written permission.*/
 #include <SDL.h>
 #include <SDL_image.h>
 #include <SDL_ttf.h>
+
+#include "Enums.h"
+#include "Bullet.h"
+#include "Structs.h"
+
+
 //#include <SpriteBox.h>f
 using namespace std;
 //Level Dimensions
@@ -44,6 +50,7 @@ const int BGTILE_SIZE = 2000;
 const int DAWN_WALK_VEL = 4;
 const int DAWN_BATTLE_VEL = 3;
 const int WALL_HEIGHT = 163;
+
 bool ControlState = false;
 bool renderflag = false;
 bool renderflag2 = false;
@@ -99,6 +106,7 @@ vector<attackboxes>  CollisionBoxArray(string FileInput){
             }
         return ReturnArray;
         }
+
 SDL_Rect BoxRectToSDL_Rect(BoxRect InRect){
     SDL_Rect OutRect;
 
@@ -115,178 +123,6 @@ SDL_Rect BoxRectToSDL_Rect(BoxRect InRect){
 // Into Action
 // and Task
 
-enum ControllerButton{
-C_UP,
-C_DOWN,
-C_LEFT,
-C_RIGHT,
-C_START,
-C_BACK,
-C_JOYLEFT,
-C_JOYRIGHT,
-C_LB,
-C_RB,
-C_A,
-C_B,
-C_X,
-C_Y
-};
-
-enum Anim{
-DAWN_WALK,
-JILL_Run,
-DAWN_PUNCH,
-DAWN_Idle,
-JILL_Dash,
-Jill_Jump,
-JILL_Shoot,
-JILL_LIGHT_KICK,
-Enemy_Idle,
-Enemy_Shoot,
-Enemy_Hit,
-MELEE_ENEMY_IDLE,
-MELEE_ENEMY_WALK,
-MELEE_ENEMY_PUNCH,
-MELEE_ENEMY_REACT,
-DAWN_PRESS_DOWN,
-DAWN_PRESS_UP,
-DAWN_PRESS_LEFT,
-DAWN_PRESS_RIGHT,
-DAWN_COMBAT_IDLE,
-DAWN_BLOCK,
-DAWN_DAMAGE,
-DAWN_CROUCH,
-DAWN_CROUCHING,
-DAWN_UNCROUCHING
-};
-
-enum CollisionLineType{
-    VERTICAL_LINE,
-    HORIZONTAL_LINE_LOWER_STOP,
-    HORIZONTAL_LINE_UPPER_STOP,
-    DIAGONAL_LINE_RIGHT_STOP,
-    DIAGONAL_LINE_LEFT_STOP,
-};
-
-enum CollisionType{
-UP_COLLISION,
-DOWN_COLLISION,
-RIGHT_COLLISION,
-LEFT_COLLISION,
-LEFT_DIAGONAL_COLLISION,
-RIGHT_DIAGONAL_COLLISION
-};
-
-
-    enum Melee_Enemy_AI_Top_State{
-    Melee_Enemy_Patrol,
-    Melee_Enemy_Attack
-
-};
-    enum Melee_Enemy_AI_Bottom_State{
-    AI_Keeping_distance_Y,
-    AI_Keeping_distance_X,
-    AI_Idle,
-    AI_Moving_In_Y,
-    AI_Moving_In_X_Attacking,
-    AI_Moving_In_Y_Attacking,
-    AI_Punch
-};
-
-    enum Melee_Enemy_AI_Move_State{
-    AI_Walk_Down,
-    AI_Walk_Up,
-    AI_No_move,
-    AI_Walk_Left,
-    AI_Walk_Right,
-    AI_Back_Left,
-    AI_Back_Right
-
-};
-struct BulletCoord{
-    float mX;
-    float mY;
-    bool flipped;
-};
-
-struct CollisionCoord{
-    float mX;
-    float mY;
-};
-struct PatrolPoint{
-
-    PatrolPoint* NextPoint;
-    int mX;
-    int mY;
-
-};
-
-class BulletHandler{
-    public:
-    BulletHandler();
-
-    // Make and destroy functions
-    // These will return true if they can succesfully create or destroy a bullet
-    bool MakeBullet(float nX, float nY, bool IsFlipped);
-    bool DestroyBullet(int index);
-    vector<BulletCoord> GetBullets();
-    BulletCoord GetBullet(int index);
-    int NumBullets();
-    void BulletX(int index, float deltaX);
-   // void BulletY(int index, int deltaY);
-    private:
-
-    vector<BulletCoord> Bulletarray;
-};
-
-BulletHandler::BulletHandler(){
-
-
-}
-
-bool BulletHandler::MakeBullet(float nx, float ny, bool Flipped){
-
-        //cout<<endl<<IsFlipped;
-        BulletCoord TempBullet{nx,ny,Flipped};
-        Bulletarray.push_back(TempBullet);
-
-        return false; //josh: had to return something
-
-
-}
-
-bool BulletHandler::DestroyBullet(int index){
-
-        Bulletarray.erase(Bulletarray.begin()+index);
-
-        return false; //josh: had to return something
-}
-
-
-vector<BulletCoord> BulletHandler::GetBullets(){
-        return Bulletarray;
-}
-
-BulletCoord BulletHandler::GetBullet(int index){
-
-        return Bulletarray.at(index);
-
-}
-
-int BulletHandler::NumBullets(){
-
-        return Bulletarray.size();
-
-}
-
-void BulletHandler::BulletX(int index, float deltaX){
-
-    BulletCoord tempCoord{Bulletarray.at(index).mX+deltaX,Bulletarray.at(index).mY,Bulletarray.at(index).flipped};
-   //cout<<endl<<"Bullet "<<index<<" has initial coord "<<Bulletarray.at(index).mX<<"it is changed to"<<tempCoord.mX;
-    Bulletarray.erase(Bulletarray.begin()+index);
-    Bulletarray.insert(Bulletarray.begin()+index,tempCoord );
-
-}
 //Texture wrapper class
 class LTexture
 {
